bootstrap: topdeps: track whether a typ is needed for export, or inline, or ...
We now keep track of what a particular typ dependency is needed for: the
prototype of a function, the body of a function, etc. For instance, if
the function is 'export inline', we print all dependencies first in the
.h.out. But if the function is only 'export' we only print the
dependencies that are required by the function prototype in the .h.out.
Types that only appear in the body can be printed in the .c.out file.

Thanks to this change, within/environment now compile without manual
edits out the output.

+++ bootstrap/types.c
@@ -19,7 +19,7 @@ struct users {
   struct users *more;
 };
 
-implement_htable_sparse(, typset, uint8_t, struct typ *);
+implement_htable_sparse(, typset, uint32_t, struct typ *);
 
 enum typ_flags {
   TYPF_TENTATIVE = 0x1,
